+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/map/type/MapDecorationType.java
@@ -1,9 +1,6 @@
 package net.mehvahdjukaar.moonlight.api.map.type;
 
-import com.mojang.serialization.Codec;
-import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.mehvahdjukaar.moonlight.api.map.CustomMapDecoration;
-import net.mehvahdjukaar.moonlight.api.map.MapDecorationRegistry;
 import net.mehvahdjukaar.moonlight.api.map.markers.MapBlockMarker;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderSet;
@@ -12,9 +9,9 @@ import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.levelgen.structure.Structure;
-import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
-
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+
 import java.util.Optional;
 
 //type itself can have 2 types: json defined or custom code defined
@@ -35,11 +32,15 @@ public abstract class MapDecorationType<D extends CustomMapDecoration, M extends
     @Nullable
     public abstract M getWorldMarkerFromWorld(BlockGetter reader, BlockPos pos);
 
-    public int getDefaultMapColor(){
+    //used for commands. gives either marker default instance or dummy marker
+    @NotNull
+    public abstract MapBlockMarker<D> getDefaultMarker(BlockPos pos);
+
+    public int getDefaultMapColor() {
         return 1;
     }
 
-    public Optional<HolderSet<Structure>> getAssociatedStructure(){
+    public Optional<HolderSet<Structure>> getAssociatedStructure() {
         return Optional.empty();
     }